move DatumVec to repr
