+++ src/dataflow/src/logging/differential.rs
@@ -26,9 +26,9 @@ use crate::activator::RcActivator;
 use crate::arrangement::manager::RowSpine;
 use crate::arrangement::KeysValsHandle;
 use crate::logging::ConsolidateBuffer;
-use crate::render::datum_vec::DatumVec;
 use crate::render::Permutation;
 use crate::replay::MzReplay;
+use repr::DatumVec;
 use repr::{Datum, Row, Timestamp};
 
 /// Constructs the logging dataflow for differential logs.
