+++ src/dataflow/src/render/join/linear_join.rs
@@ -26,9 +26,9 @@ use repr::{Diff, Row, RowArena};
 use crate::operator::CollectionExt;
 use crate::render::context::CollectionBundle;
 use crate::render::context::{Arrangement, ArrangementFlavor, ArrangementImport, Context};
-use crate::render::datum_vec::DatumVec;
 use crate::render::join::{JoinBuildState, JoinClosure};
 use crate::render::Permutation;
+use repr::DatumVec;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct LinearJoinPlan {
