+++ src/dataflow/src/render/upsert.rs
@@ -244,7 +244,7 @@ where
     // If we have temporal predicates do the thing they have to do.
     if let Some(plan) = temporal_plan {
         let (oks2, errs2) = oks.flat_map_fallible("UpsertTemporalOperators", {
-            let mut datums = crate::render::datum_vec::DatumVec::new();
+            let mut datums = repr::DatumVec::new();
             let mut row_builder = Row::default();
             move |(row, time, diff)| {
                 let arena = repr::RowArena::new();
