+++ src/repr/src/lib.rs
@@ -22,6 +22,7 @@
 
 #![warn(missing_debug_implementations)]
 
+mod datum_vec;
 mod relation;
 mod row;
 mod scalar;
@@ -30,6 +31,7 @@ pub mod adt;
 pub mod strconv;
 pub mod util;
 
+pub use datum_vec::{DatumVec, DatumVecBorrow};
 pub use relation::{ColumnName, ColumnType, NotNullViolation, RelationDesc, RelationType};
 pub use row::{
     datum_list_size, datum_size, datums_size, row_size, DatumList, DatumMap, Row, RowArena, RowRef,
